src/app/classes/DataSetFilter.spec.ts
import { Params } from '@angular/router';

import { DataGrouping, FilterRule } from '../interfaces/WidgetInterfaces';
import { DataSet } from '../interfaces/DisplayInterfaces';

import { DataSetFilter } from './DataSetFilter';


describe('DataSetFilter.FilterDataSet', () => {
    const dataSet: DataSet = getDataSet();

    it('should produce the record count expected for context filter', () => {
        const context: Params = [];
        const rules: FilterRule[] = [];
        rules.push({column: 'Design', contextValueName: 'Design', serverSide: false });
        context['Design'] = 'Z21D';
        const result = DataSetFilter.FilterDataSet(dataSet, rules, context, undefined);
        expect(result.data.length).toBe(41);

    });

    it('should produce the record count expected cor context and regular filter', () => {
        const context: Params = [];
        const filter: Params = [];
        const rules: FilterRule[] = [];
        rules.push({column: 'Design', contextValueName: 'Design', serverSide: false });
        context['Design'] = 'Z21D';
        filter['Step'] = 'ncstartswith|q';
        const result = DataSetFilter.FilterDataSet(dataSet, rules, context, filter);
        expect(result.data.length).toBe(6);

    });

});

describe('DataSetFilter.GroupDataSet', () => {
const ds = getNumericDataSet();
    const dg: DataGrouping = {
        groupBy: ['Facility'],
        aggregations: [
            { name: 'Wafer', operation: 'mean', rename: 'w mean' }
        ]
    };

    it('should aggregate mean correctly', () => {
        console.log(JSON.stringify(dg));
        const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
        console.log(JSON.stringify(groupedDs));
        expect(groupedDs.data[0][1]).toBeCloseTo(6853.36);
    });

    it('should aggregate count correctly', () => {
        dg.aggregations[0].operation = 'count';
        console.log(JSON.stringify(dg));
        const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
        console.log(JSON.stringify(groupedDs));
        expect(groupedDs.data[0][1]).toBeCloseTo(55);
    });

    it('should aggregate median correctly', () => {
        dg.aggregations[0].operation = 'median';
        console.log(JSON.stringify(dg));
        const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
        console.log(JSON.stringify(groupedDs));
        expect(groupedDs.data[0][1]).toBeCloseTo(7623);
    });


    it('should aggregate min correctly', () => {
        dg.aggregations[0].operation = 'min';
        console.log(JSON.stringify(dg));
        const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
        console.log(JSON.stringify(groupedDs));
        expect(groupedDs.data[0][1]).toBeCloseTo(123);
    });

    it('should be able to handle bad column data', () => {
        dg.aggregations[0].operation = 'stdDev';
        ds.data[2][5] = 'not a number';
        console.log(JSON.stringify(dg));
        const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
        console.log(JSON.stringify(groupedDs));
        expect(groupedDs.data[0][1]).toBeCloseTo(2807.85);
    });

    it('should handle specs with non-existent columns correctly', () => {
        dg.groupBy.push('InvalidColumn');

        dg.aggregations.push({ name: 'Invalid Agg Column', operation: 'stdDev', rename: 'doesnt matter' });
        console.log(JSON.stringify(dg));
        const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
        console.log(JSON.stringify(groupedDs));
        expect(groupedDs.data[0][1]).toBeCloseTo(2807.85);
    });

    it('should error when an aggregation type does not exist', () => {
        dg.aggregations[0].operation = 'bogus';
        console.log(JSON.stringify(dg));
        expect(function(){  DataSetFilter.GroupDataSet(ds, dg); }).toThrowError();
    });

    it ('should do no aggregations when no aggregations are supplied', () => {
       dg.aggregations = [];
       dg.groupBy = ['Facility'];
       console.log(JSON.stringify(dg));
       const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
       expect(groupedDs.columns.length).toBe(1);
    });
    // it('should aggregate mean correctly with a bad column value', () => {

    //     ds.data[2][5] = 'not a number';
    //     dg.aggregations[0].value = 'max';
    //     console.log(JSON.stringify(dg, null, 2));
    //     const groupedDs = DataSetFilter.GroupDataSet(ds, dg);
    //     console.log(JSON.stringify(groupedDs));
    //     expect(groupedDs).toBeDefined();
    // });


});

function getNumericDataSet(): DataSet {
    const ds = getDataSet();

    ds.data.forEach( row => {
        row[5] = row[5].split('-')[0];
    });


    return ds;

}

function getDataSet(): DataSet {
    return {
        columns : ['Facility', 'Design', 'Step', 'Lot', 'StartLot', 'Wafer'],
        data : [
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-07'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-03'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-06'],
            ['fab4', 'R80M', 'HQB1WP.00', '3456563.0C3', '3456563', '6563-13'],
            ['fab4', 'R80M', 'HQR1PP.00', '3456563.0R3', '3456563', '6563-02'],
            ['fab4', 'Z21D', 'Q1RPP.00', '3460123.003', '3460123', '0123-12'],
            ['fab4', 'Z21D', 'RPC.00', '3477103.1K3', '3477103', '7103-02'],
            ['fab2', 'S15C', 'UCYPP.00', '9927522.003', '9927522', '7522-06'],
            ['fab2', 'S15C', 'CZPP.00', '9935982.003', '9935982', '5982-01'],
            ['fab2', 'S15C', 'UCYPP.00', '9951512.003', '9951512', '1512-15'],
            ['fab10', 'B16A', 'RCB3WP.00', '8317081.023', '8317081', '7081-19'],
            ['fab2', 'S15C', 'UCYPP.00', '9927522.003', '9927522', '7522-10'],
            ['fab4', 'R80M', 'CQPP.00', '3456563.0H3', '3456563', '6563-05'],
            ['fab4', 'R80M', 'HDR2PP.00', '3463503.0B3', '3463503', '3503-24'],
            ['fab4', 'Z21D', 'Q1RPP.00', '3460123.003', '3460123', '0123-14'],
            ['fab4', 'Z21D', 'Q1RPP.00', '3460123.003', '3460123', '0123-16'],
            ['fab2', 'S15C', 'CZPP.00', '9935582.003', '9935582', '5582-25'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-14'],
            ['fab2', 'S26A', 'WCPC.00', '9874452.003', '9874452', '4452-12'],
            ['fab2', 'S15C', 'UCYPP.00', '9912312.003', '9912312', '2312-25'],
            ['fab2', 'S15C', 'UCYPP.00', '9951512.003', '9951512', '1512-13'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-16'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-17'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-18'],
            ['fab4', 'R80M', 'HQR2PP.00', '3456563.0C3', '3456563', '6563-13'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-16'],
            ['fab4', 'R80M', 'UHDR1PC.00', '3460383.0W3', '3460383', '0383-20'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-02'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-03'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-10'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-11'],
            ['fab2', 'S15C', 'RDR1PP.00', '9935982.003', '9935982', '5982-16'],
            ['fab2', 'S15C', 'CZPP.00', '9943592.013', '9943592', '3592-23'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9927522.003', '9927522', '7522-06'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9927522.003', '9927522', '7522-08'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9927522.003', '9927522', '7522-20'],
            ['fab2', 'S26A', 'FPC.00', '9878702.003', '9878702', '8702-08'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9950872.013', '9950872', '0872-10'],
            ['fab2', 'S15C', 'CZPP.00', '9935982.003', '9935982', '5982-03'],
            ['fab2', 'S15C', 'CZPP.00', '9943592.013', '9943592', '3592-01'],
            ['fab2', 'S15C', 'CZPP.00', '9943592.003', '9943592', '3592-21'],
            ['fab2', 'S15C', 'CZPP.00', '9950902.003', '9950902', '0902-19'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9951512.003', '9951512', '1512-10'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9951512.003', '9951512', '1512-21'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-03'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-05'],
            ['fab4', 'R80M', 'HQB1WP.00', '3456563.0H3', '3456563', '6563-05'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-03'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-04'],
            ['fab4', 'R80M', 'CYPP.00', '3409053.10J', '3409053', '9053-10'],
            ['fab4', 'R80M', 'HQB1WP.00', '3456563.0R3', '3456563', '6563-02'],
            ['fab4', 'Z21D', 'Q1RPP.00', '3460123.003', '3460123', '0123-11'],
            ['fab4', 'Z21D', 'CPC.00', '3439423.003', '3439423', '9423-09'],
            ['fab4', 'Z21D', 'RDPP.00', '3439423.003', '3439423', '9423-09'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-16'],
            ['fab4', 'Z21D', 'C2PC.00', '3439423.003', '3439423', '9423-09'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-11'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-02'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-13'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-03'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-15'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9927522.003', '9927522', '7522-06'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9927522.003', '9927522', '7522-08'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9927522.003', '9927522', '7522-10'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9927522.003', '9927522', '7522-20'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9927522.003', '9927522', '7522-25'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9951512.003', '9951512', '1512-10'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9951512.003', '9951512', '1512-21'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9951512.003', '9951512', '1512-23'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-02'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-15'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-15'],
            ['fab2', 'S15C', 'UCYPP.00', '9927522.003', '9927522', '7522-04'],
            ['fab2', 'S15C', 'CZPP.00', '9935632.003', '9935632', '5632-08'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-07'],
            ['fab10', 'B27A', 'FPP.00', '8299261.003', '8299261', '9261-02'],
            ['fab10', 'N18A', 'XDR2PC.00', '8357101.003', '8357101', '7101-15'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-24'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9912312.003', '9912312', '2312-25'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9951512.003', '9951512', '1512-13'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9951512.003', '9951512', '1512-23'],
            ['fab4', 'R80M', 'HQB2WP.00', '3456563.0C3', '3456563', '6563-13'],
            ['fab2', 'S26A', 'WCPC.00', '9869852.0B3', '9869852', '9852-06'],
            ['fab2', 'S26A', 'WCPC.00', '9869852.0B3', '9869852', '9852-06'],
            ['fab2', 'S26A', 'WCPC.00', '9869852.0B3', '9869852', '9852-06'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9912312.003', '9912312', '2312-20'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9912312.003', '9912312', '2312-23'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9927522.003', '9927522', '7522-04'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9927522.003', '9927522', '7522-10'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9927522.003', '9927522', '7522-25'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9951512.003', '9951512', '1512-15'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9951512.003', '9951512', '1512-16'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-19'],
            ['fab2', 'S15C', 'CZPP.00', '9935582.003', '9935582', '5582-16'],
            ['fab2', 'S15C', 'UCYPP.00', '9950872.013', '9950872', '0872-10'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-02'],
            ['fab2', 'S15C', 'CZPP.00', '9935622.003', '9935622', '5622-18'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-01'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935982.003', '9935982', '5982-25'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-12'],
            ['fab4', 'R80M', 'HQR1PP.00', '3456563.0C3', '3456563', '6563-13'],
            ['fab4', 'R80M', 'CQPP.00', '3456563.0R3', '3456563', '6563-02'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-20'],
            ['fab2', 'S15C', 'UCYPP.00', '9927522.003', '9927522', '7522-08'],
            ['fab4', 'R80M', 'HQR1PP.00', '3456563.0H3', '3456563', '6563-05'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-16'],
            ['fab4', 'Z21D', 'Q1RPP.00', '3460123.003', '3460123', '0123-13'],
            ['fab4', 'Z21D', 'RPP.00', '3468163.003', '3468163', '8163-05'],
            ['fab4', 'Z21D', 'Q1RPP.00', '3460123.003', '3460123', '0123-15'],
            ['fab4', 'Z21D', 'RPP.00', '3468163.003', '3468163', '8163-02'],
            ['fab10', 'B16A', 'RQR3PE.00', '8317081.023', '8317081', '7081-19'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-01'],
            ['fab4', 'Z21D', 'RPP.00', '3468163.003', '3468163', '8163-04'],
            ['fab2', 'S15C', 'UCYPP.00', '9912312.003', '9912312', '2312-20'],
            ['fab2', 'S15C', 'UCYPP.00', '9951512.003', '9951512', '1512-16'],
            ['fab10', 'B27A', 'FPP.00', '8299261.1G3', '8299261', '9261-03'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-09'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-12'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-18'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-23'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-12'],
            ['fab4', 'Z21D', 'RCPC.00', '3439423.003', '3439423', '9423-09'],
            ['fab2', 'S15C', 'CZPP.00', '9950902.003', '9950902', '0902-17'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-04'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-13'],
            ['fab2', 'S15C', 'UCYPP.00', '9951512.003', '9951512', '1512-10'],
            ['fab2', 'S26A', 'WCPC.00', '9869852.0B3', '9869852', '9852-01'],
            ['fab2', 'S26A', 'WCPC.00', '9869852.0B3', '9869852', '9852-01'],
            ['fab2', 'S15C', 'CZPC.00', '9860182.1T3', '9860182', '0182-24'],
            ['fab2', 'S15C', 'CZPP.00', '9943592.003', '9943592', '3592-25'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-08'],
            ['fab2', 'S15C', 'RDR1PP.00', '9927522.003', '9927522', '7522-08'],
            ['fab2', 'S15C', 'UCYPP.00', '9912312.003', '9912312', '2312-23'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-17'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9912312.003', '9912312', '2312-20'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-04'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-11'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-17'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-21'],
            ['fab2', 'S15C', 'UCYPP.00', '9927522.003', '9927522', '7522-25'],
            ['fab2', 'S15C', 'CZPP.00', '9950902.003', '9950902', '0902-06'],
            ['fab2', 'S15C', 'UCYPP.00', '9951512.003', '9951512', '1512-23'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-05'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9912312.003', '9912312', '2312-23'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9912312.003', '9912312', '2312-25'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9927522.003', '9927522', '7522-04'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9951512.003', '9951512', '1512-13'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9951512.003', '9951512', '1512-15'],
            ['fab2', 'S15C', 'UHDB1WP.00', '9951512.003', '9951512', '1512-16'],
            ['fab4', 'Z21D', 'RPP.00', '3437623.003', '3437623', '7623-10'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-02'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-07'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-14'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-13'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-03'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-04'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-06'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-11'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-17'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-18'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-19'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-22'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-23'],
            ['fab2', 'S15C', 'RCB1WP.00', '9950902.003', '9950902', '0902-25'],
            ['fab2', 'S15C', 'UHDR1PP.00', '9950872.013', '9950872', '0872-10'],
            ['fab2', 'S15C', 'CZPP.00', '9951552.003', '9951552', '1552-19'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-10'],
            ['fab2', 'S15C', 'UCYPP.00', '9927522.003', '9927522', '7522-20'],
            ['fab2', 'S15C', 'UCYPP.00', '9951512.003', '9951512', '1512-21'],
            ['fab10', 'L28Y', 'FPP.00', '8319251.003', '8319251', '9251-06'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-05'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-10'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-11'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-12'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-16'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-22'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-25'],
            ['fab2', 'S15C', 'RCB1WP.00', '9935582.003', '9935582', '5582-23'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-12'],
            ['fab2', 'S26A', 'FPP.00', '9783782.003', '9783782', '3782-21'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-06'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-08'],
            ['fab2', 'S15C', 'CZPP.00', '9935632.003', '9935632', '5632-19'],
            ['fab2', 'S15C', 'RDR1PP.00', '9935982.003', '9935982', '5982-10'],
            ['fab4', 'Z21D', 'RPP.00', '3437653.003', '3437653', '7653-11'],
            ['fab4', 'Z21D', 'RPC.00', '3439423.003', '3439423', '9423-09'],
            ['fab4', 'R80M', 'CQPP.00', '3456563.0D3', '3456563', '6563-03'],
            ['fab4', 'Z21D', 'RPP.00', '3469113.003', '3469113', '9113-08'],
            ['fab4', 'Z21D', 'RPP.00', '3468163.003', '3468163', '8163-03'],
            ['fab10', 'B27A', 'FPP.00', '8244161.003', '8244161', '4161-11']
        ]
    };
}
